test: enable high-verbosity logging in every component test config (#316)
None of the component test configs declared a logger: component,
which makes ESPHome silently compile ESP_LOGCONFIG/ESP_LOGW down to
no-ops and skip type-checking their arguments entirely (see #315,
where this exact gap let a broken ESPBTUUID::to_str() call through
CI). Add logger: level: VERY_VERBOSE to every component's test config
(via the shared common.yaml where one exists) so dump_config() and
warning-log code paths actually get compiled and type-checked.

Recompiled every component's esp32-idf/esp32-ard test config with
this change and fixed the one warning it caught: a %u format
specifier vs. uint32_t argument mismatch in divoom_display.cpp's
dump_config(), the same class of bug already fixed in ble_elm327.

+++ components/divoom/divoom_display.cpp
@@ -20,7 +20,7 @@ void DivoomDisplay::dump_config()
     ESP_LOGCONFIG(TAG, "  MAC address        : %s", this->parent_->address_str());
     ESP_LOGCONFIG(TAG, "  Service UUID       : %s", this->service_uuid_.to_str(uuid_buf));
     ESP_LOGCONFIG(TAG, "  Characteristic UUID: %s", this->char_uuid_.to_str(uuid_buf));
-    ESP_LOGCONFIG(TAG, "  Update Interval: %u ms", this->get_update_interval());
+    ESP_LOGCONFIG(TAG, "  Update Interval: %lu ms", (unsigned long) this->get_update_interval());
 }
 
 void DivoomDisplay::update()
